initial commit
This is a Git reproduction of history of changes of userscript "Twitter:
bring back old name and logo".  Source code of the userscript was stored
only on Greasy Fork.

This is a reproduction of the first version of the script, which was
called just "Twitter: old icon" back then.  URL to the version is:

https://greasyfork.org/en/scripts/471871-twitter-bring-back-old-name-and-logo/code?version=1226584