renameHiddenTweets: rename hidden quoted tweets
The selector in renameHiddenTweets is targeted to hidden tweets in
replies.  Add additional selector to target quoted tweets at the top of
the opened thread.  Such hidden quoted tweets can have messages like
"This Tweet is from an account that no longer exists."

Sidenote: the renaming of replies doesn't work very well.  It only works
on full reload of a tweet's page.  It doesn't work if user navigates
from the timeline to a singular tweets.