avoid waiting for elements multiple times
`MutationObserver`s from calls to waitForElement multiply due to many
calls to rename() and other top-level functions.  Replace library
function waitForElement with uniqueWaitForElement.  Keep track of the
observers based on the selectors, which now have to be unique -- a
selector cannot be used multiple times with different purposes.

Example of the requirement of uniqueness of the observers -- rewrite
renameTweetPlaceholders to only use a single observer instead of an
observer for every combination of target and replacement texts.

This should, in general, give a performance improvement due to
dramatically smaller amount of `MutationObserver`s used on a page.

Note: parameter needReplacing in new function uniqueWaitForElement is
not used yet.  It was an idea that didn't pan out.