*: overhaul selectors
Recent updates on Twitter changed CSS classes, which are used by the
userscript.  This has also affected userstyles for Twitter that I
develop.  See:

   - v4 of https://userstyles.world/style/10864/twitter-hide-trends-and-who-to-follow
   - v6 of https://userstyles.world/style/12377/twitter-hide-tab-for-you

Update the affected selectors.  To test, I went by the list on Greasy
Fork of the things that the userscript is supposed to change.

	https://greasyfork.org/en/scripts/471871-twitter-bring-back-old-name-and-logo

I might have missed some, but that probably means that it's something
not easily visible.  Keep in mind the list in the comment at the top of
the code, titled "Things deliberately left with the new name".